// oppabControlliClient/src/app/model/checklistRegola.ts
import { Controllo } from './controllo';

export class ChecklistRegola {

  id: number;
  idChecklistRegola: number;
  controllo: Controllo;
  anno: number;
  campagna: number;
  misura: string;
  intervento: string;
  sottoIntervento: string;
  tipoCampione: string;
  tipoControllo: string;
  idChecklistTemplate: number;

}